PhantomStrike_AI.py
import logging
import signal
import sys
import asyncio
import datetime
import time
import re
import json
import requests
import socket
import base64
import yaml
import threading
import smtplib
try:
	from cryptography.fernet import Fernet
except ImportError:
	Fernet = None
try:
	from fastapi import FastAPI
	import uvicorn
except ImportError:
	FastAPI = None
	uvicorn = None

from modules import utils, validation, exploits, scanner, analyzer
# from modules import reporting  # Da implementare se necessario


# Placeholder per funzioni/oggetti non definiti
def get_opt(key, default=None):
	return default

class CredentialVault:
	def __init__(self, mode="memory", filename=None, key=None):
		self.mode = mode
		self.filename = filename
		self.key = key
	def get(self, target, service):
		return {"username": "anonymous", "password": "test@test.com"}
	def set(self, tgt, svc, user, pwd):
		pass

def _db_execute(*args, **kwargs):
	pass

class RedTeamBotCore:
	pass

class PluginManager:
	def __init__(self, core, plugins_dir="plugins"):
		pass

def _signal_handler(sig, frame):
	try:
		close_session(globals().get("_current_session_id", None), 0, 0, 0)
		logger.info("Sessione DB chiusa correttamente.")
	except Exception:
		pass
	sys.exit(130)

logger = logging.getLogger("PhantomStrike AI")
if not logger.hasHandlers():
	logging.basicConfig(level=logging.INFO)

signal.signal(signal.SIGINT, _signal_handler)

def close_session(session_id, total_ports, total_vulns, total_success, duration=0.0):
	pass


def save_result(session_id, target, port, service, banner, vuln, severity, ai_score, exploit_success, details):
	_db_execute(
		"""INSERT INTO results
			(session_id, target, port, service, banner, vuln, severity, ai_score, exploit_success, details, timestamp)
			VALUES (?,?,?,?,?,?,?,?,?,?,?)""",
		(session_id, target, port, service, banner, vuln, severity, ai_score,
		 int(exploit_success), details, datetime.datetime.now().isoformat())
	)


# =========================
# All scan, analysis, and exploit functions are now in modules


# =========================
# All report and export functions are now in modules


# =========================
# The entire main engine now uses modules


# =========================
# CLI ARGUMENT PARSER (to be completed if necessary)
def parse_args():
	pass


# =========================
# ENTRY POINT
def start_api(bot_core, plugin_manager):
	if FastAPI is None or uvicorn is None:
		logger.error("FastAPI/uvicorn not installed. Install with: pip install fastapi uvicorn")
		return
	app = FastAPI(title="PhantomStrike AI API", description="Red Team scan orchestration and automation", version="1.0.0")

	scans = {}
	scan_counter = [0]

	@app.post("/scan")
	async def scan_api(body: dict):
		targets = body.get("targets")
		options = body.get("options", {})
		if not targets or not isinstance(targets, list):
			raise Exception("Field 'targets' is required and must be a list")
		scan_id = scan_counter[0] = scan_counter[0] + 1
		scans[scan_id] = {"status": "in_progress", "targets": targets, "results": None}

		def run_scan():
			# Runs synchronous scan for demo simplicity
			results = []
			for tgt in targets:
				try:
					open_ports = asyncio.run(scanner.scan_target(tgt))
					vulns = analyzer.analyze_ports(tgt, open_ports)
					results.append({"target": tgt, "vulns": vulns})
				except Exception as e:
					results.append({"target": tgt, "error": str(e)})
			scans[scan_id]["status"] = "done"
			scans[scan_id]["results"] = results

		threading.Thread(target=run_scan, daemon=True).start()
		return {"scan_id": scan_id, "status": "in_progress"}

	@app.get("/status")
	def status():
		return {"scans": {k: v["status"] for k, v in scans.items()}}

	@app.get("/results/{scan_id}")
	def results(scan_id: int):
		if scan_id not in scans:
			from fastapi.exceptions import HTTPException as get_http_exception
			raise get_http_exception(status_code=404, detail="Scan ID not found")
		return scans[scan_id]

	# Start uvicorn server (localhost only)
	uvicorn.run(app, host="127.0.0.1", port=8000, log_level="info")

if __name__ == "__main__":
	print(f"""
# Banner in English
	{utils.C.CYAN}{utils.C.BOLD}╔══════════════════════════════════════════════════════════╗
	║       PhantomStrike AI — Enterprise Pen Test Assistant   ║
	║       ⚠️  FOR AUTHORIZED/ LAB ENVIRONMENTS ONLY         ║
	╚══════════════════════════════════════════════════════════╝{utils.C.RESET}
	""")
# PhantomStrike AI

#This file has been renamed from ai_redteam_bot.py to PhantomStrike_AI.py as requested. All main references have been updated. Make sure to also update any external references or automation scripts pointing to the old name.