requirements.txt
requests>=2.31
pyyaml>=6.0
reportlab>=4.0
pytest>=7.0
uvicorn>=0.23
fastapi>=0.110
# stdlib: logging, signal, sys, asyncio, datetime, time, re, json, socket, base64, threading, smtplib, ftplib, os
# note: ftplib, smtplib, socket, threading, base64, logging, etc. are part of Python standard library